Viele neuerungen, nur an meinen und allgemeinen Klassen (persistence etc) Einige Klassen von euch werden nun Fehlerhaft sein, muss noch angepasst werden. Mehr dazu im ToDo

+++ shop/src/main/java/de/shop/util/Constants.java
@@ -2,7 +2,7 @@ package de.shop.util;
 
 public final class Constants {
 	// JPA
-	public static final Long KEINE_ID = null;
+	public static final Integer KEINE_ID = null;
 	public static final long MIN_ID = 1L;
 	public static final int ERSTE_VERSION = 0;
